+++ lib/aws_iot_device.dart
@@ -1,3 +1,5 @@
 export './src/device.dart';
+export './src/signed_request.dart';
+export './src/socket_connection.dart';
 export 'package:tuple/tuple.dart';
 export 'package:amazon_cognito_identity_dart/sig_v4.dart';