+++ lib/src/socket_connection.dart
@@ -1,0 +1,7 @@
+library connection;
+
+import 'package:mqtt/mqtt_shared.dart';
+
+abstract class SocketConnection {
+	VirtualMqttConnection getConnection(String url);
+}