+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+- BREAKING CHANGE: Switched out MQTT libary to one that supports HTML Web Sockets
+- This means the library API is different
+
 ## 0.0.4
 
 - Added missing "unsubscribe" method